Update instancing.rst
In version 2.1 of Godot the "+" button is not for creating instances. The chain-link button is the button for instances, so when following the tutorial I was lost. Please correct the tutorial with the corrected text "chain link" instead of "+" and include a new image that shows the updated control panel and the chain link button.